+++ develappscleanarchitecture/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [1.0.3]
+
+- Add injection container to the generated screen.
+- Add Multibox to the generated screen.
+
 ## [1.0.1]
 
 - Rename pages to screens